requirements.txt
easyocr
manga-ocr
paddlepaddle-gpu>=2.6.0,<3.0.0
paddleocr
deep_translator
Pillow>=10.0
wget
colorama
nest_asyncio
nest-asyncio
python-dotenv
pydrive2
groq
loguru